move ParamValueSpinner to a separate file
